requirements/common.txt
alembic==1.4.3
appdirs==1.4.4
appnope==0.1.2
APScheduler==3.6.3
attrs==20.3.0
Automat==20.2.0
backcall==0.2.0
blinker==1.4
Bootstrap-Flask==1.5.1
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
constantly==15.1.0
cryptography==3.3
cssselect==1.1.0
decorator==4.4.2
demjson==2.2.4
dnspython==2.0.0
elasticsearch==7.10.0
email-validator==1.1.2
Flask==1.1.2
Flask-Admin==1.5.7
Flask-APScheduler==1.11.0
Flask-Cors==3.0.10
Flask-HTTPAuth==4.2.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
flask-shell-ipython==0.4.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
itemadapter==0.2.0
itemloaders==1.0.4
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
jupyter-client==6.1.7
jupyter-core==4.7.0
lxml==4.6.2
Mako==1.1.3
MarkupSafe==1.1.1
parsel==1.6.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
Protego==0.1.16
psycopg2==2.8.6
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
Pygments==2.7.3
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==20.0.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-telegram-bot==13.4.1
pytz==2020.4
pyzmq==20.0.0
queuelib==1.5.0
requests==2.25.0
Scrapy==2.4.1
scrapyrt==0.11.0
service-identity==18.1.0
six==1.15.0
SQLAlchemy==1.3.20
telegram-send==0.25
tornado==6.1
traitlets==5.0.5
Twisted==20.3.0
tzlocal==2.1
Unidecode==1.2.0
urllib3==1.26.2
w3lib==1.22.0
wcwidth==0.2.5
Werkzeug==1.0.1
whitenoise==5.2.0
WTForms==2.3.3
zope.interface==5.2.0
